+++ Converter/src/indexer.cpp
@@ -22,7 +22,7 @@ namespace indexer{
 		double x;
 		double y;
 		double z;
-		int32_t pointIndex;
+		int64_t pointIndex;
 		int32_t childIndex;
 	};
